+++ lib/d_input.dart
@@ -1,6 +1,9 @@
+library d_input;
 import 'package:flutter/material.dart';
 
 typedef DValidator = String? Function(String? input);
+typedef OnChange = void Function(String? value);
+typedef OnTap = void Function();
 
 /// simple input with full border default
 /// to styling the input decoration, use theme in the main
@@ -16,6 +19,9 @@ class DInput extends StatelessWidget {
     this.inputType,
     this.maxLine,
     this.minLine,
+    this.onTap,
+    this.onChanged,
+    this.spaceTitle,
   }) : super(key: key);
 
   /// controller for input
@@ -40,6 +46,12 @@ class DInput extends StatelessWidget {
   /// to activate validator, wrap input with 'Form widget'
   final DValidator? validator;
 
+  /// get realtime value from input
+  final OnChange? onChanged;
+
+  /// action tap input
+  final OnTap? onTap;
+
   /// type for the input, like number
   final TextInputType? inputType;
 
@@ -48,6 +60,10 @@ class DInput extends StatelessWidget {
 
   /// minimum line for input
   final int? minLine;
+
+  /// space beetween input and title text
+  /// default: 8
+  final double? spaceTitle;
   @override
   Widget build(BuildContext context) {
     return Column(
@@ -64,13 +80,13 @@ class DInput extends StatelessWidget {
               ),
               if (isRequired ?? false) const SizedBox(width: 2),
               if (isRequired ?? false)
-                Text(
+                const Text(
                   '*',
-                  style: TextStyle(color: Theme.of(context).colorScheme.error),
+                  style: TextStyle(color: Colors.red),
                 ),
             ],
           ),
-        if (title != null) const SizedBox(height: 4),
+        if (title != null) SizedBox(height: spaceTitle ?? 8),
         TextFormField(
           controller: controller,
           validator: validator,
@@ -78,6 +94,8 @@ class DInput extends StatelessWidget {
           keyboardType: inputType,
           minLines: minLine ?? 1,
           maxLines: maxLine ?? 1,
+          onTap: onTap,
+          onChanged: onChanged,
           decoration: InputDecoration(
             border: const OutlineInputBorder(),
             isDense: true,
@@ -92,10 +110,149 @@ class DInput extends StatelessWidget {
                           Text(label!),
                           if (isRequired ?? false) const SizedBox(width: 2),
                           if (isRequired ?? false)
-                            Text(
+                            const Text(
                               '*',
-                              style: TextStyle(
-                                  color: Theme.of(context).colorScheme.error),
+                              style: TextStyle(color: Colors.red),
+                            ),
+                        ],
+                      ),
+          ),
+        ),
+      ],
+    );
+  }
+}
+
+/// simple input password with full border default
+/// to styling the input decoration, use theme in the main
+class DInputPassword extends StatefulWidget {
+  const DInputPassword({
+    Key? key,
+    required this.controller,
+    this.label,
+    this.title,
+    this.isRequired,
+    this.hint,
+    this.validator,
+    this.inputType,
+    this.maxLine,
+    this.minLine,
+    this.onTap,
+    this.onChanged,
+    this.spaceTitle,
+    this.obsecureCharacter,
+  }) : super(key: key);
+
+  /// controller for input
+  final TextEditingController controller;
+
+  /// text inline border input
+  /// if [title] not null, label will be replace
+  /// [label] < [title]
+  final String? label;
+
+  /// [title] will replace [label], if label not null
+  /// [title] > [label]
+  final String? title;
+
+  /// dummy placeholder
+  final String? hint;
+
+  /// asteris symbol will shown if required
+  final bool? isRequired;
+
+  /// input validator
+  /// to activate validator, wrap input with 'Form widget'
+  final DValidator? validator;
+
+  /// get realtime value from input
+  final OnChange? onChanged;
+
+  /// action tap input
+  final OnTap? onTap;
+
+  /// type for the input, like number
+  final TextInputType? inputType;
+
+  /// maximum line for input
+  final int? maxLine;
+
+  /// minimum line for input
+  final int? minLine;
+
+  /// space beetween input and title text
+  /// default: 8
+  final double? spaceTitle;
+
+  // character which show when obsecure is active to hide real character
+  final String? obsecureCharacter;
+
+  @override
+  State<DInputPassword> createState() => _DInputPasswordState();
+}
+
+class _DInputPasswordState extends State<DInputPassword> {
+  bool isObsecure = true;
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      mainAxisSize: MainAxisSize.min,
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        if (widget.title != null)
+          Row(
+            mainAxisSize: MainAxisSize.min,
+            children: [
+              Text(
+                widget.title!,
+                style: const TextStyle(fontWeight: FontWeight.bold),
+              ),
+              if (widget.isRequired ?? false) const SizedBox(width: 2),
+              if (widget.isRequired ?? false)
+                const Text(
+                  '*',
+                  style: TextStyle(color: Colors.red),
+                ),
+            ],
+          ),
+        if (widget.title != null) SizedBox(height: widget.spaceTitle ?? 8),
+        TextFormField(
+          controller: widget.controller,
+          obscureText: isObsecure,
+          validator: widget.validator,
+          autovalidateMode: AutovalidateMode.onUserInteraction,
+          keyboardType: widget.inputType,
+          minLines: widget.minLine ?? 1,
+          maxLines: widget.maxLine ?? 1,
+          onTap: widget.onTap,
+          onChanged: widget.onChanged,
+          obscuringCharacter: widget.obsecureCharacter ?? '●',
+          decoration: InputDecoration(
+            suffixIcon: IconButton(
+              onPressed: () {
+                setState(() {
+                  isObsecure = !isObsecure;
+                });
+              },
+              icon: Icon(isObsecure ? Icons.visibility_off : Icons.visibility),
+            ),
+            border: const OutlineInputBorder(),
+            isDense: true,
+            hintText: widget.hint,
+            label: widget.title != null
+                ? null
+                : widget.label == null
+                    ? null
+                    : Row(
+                        mainAxisSize: MainAxisSize.min,
+                        children: [
+                          Text(widget.label!),
+                          if (widget.isRequired ?? false)
+                            const SizedBox(width: 2),
+                          if (widget.isRequired ?? false)
+                            const Text(
+                              '*',
+                              style: TextStyle(color: Colors.red),
                             ),
                         ],
                       ),