+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.0.7
+
+- add widget input password
+
+## 0.0.6
+
+- add properti onChange
+- add properti onTap
+- add properti spaceTitle & change default spaceTitle
+- change color required to const Red
+
 ## 0.0.2
 
 - add tutorial